requirements.txt
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
requests==2.25.1
beautifulsoup4==4.9.3
xlwt==1.3.0
jwt==1.2.0